Integrate darwintab
Modularized the GA module, the pitch estimation, and transcription
modules so they can be used independantly. TODOs have been placed in the
code where things need to be changed to make things a little neater.
I'm in a rush to push out something to show for thesis meeting, but it
works.