Write meitab to media location
Instead of writing to a tmp file in darwin tab, re-opening the file in
robotaba and passing it to the FileField, darwin tab can now return a
string representation of the Mei Tab, which is simply written to the
media_url location specified by the FileField in the MeiTab model.
Should not run into any problems with MeiDocument strings exceeding the
size of main memory, so this is a good solution.

Fixes #4