Encode Guitar MetaData
Capo and string tuning in MEI file. MEI has no way to encode capo
position like MusicXML does, without a custom build of MEI that is. In
MEI the capo could be implicitly encoded in the string tuning attribute,
but it's ambiguous whether the strings should be tuned that way or
whether a capo should be applied.